LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigation-history-entry/entry-after-detach-expected.txt


FAIL NavigationHistoryEntry properties after detach assert_equals: expected (string) "" but got (object) null
